Simplify layout to not use absolute positions
This avoids the problem of text overlapping each other.

The toolbar is still on the top right as before, but now with
`float: right;` instead of `position: absolute;`. This was done
by making the header `float: left;`, and hiding the element in-between
(the pass/fail banner).

The banner was re-implemented as the top border of #qunit-userAgent.