Minor design improvements to accomodate QUnit 2.18.2
This version still works fine with older versions of QUnit as well,
but it now handles better the changes to the module filter buttons
in QUnit 2.18.2.

Also fix a few old alignment bugs, and make it not pointer-cursor
hover the unclickable `<strong>` element for test source URL. This
style was intended for the (clickable) test name, which happens to
also be in a `<strong>` element. Not great, to be improved in QUnit 3
with new HTML Reporter structure.

Ref https://github.com/qunitjs/qunit/issues/1301.